In safe struct wrappers, do not use Bool8.
On .NET Core Unix platforms, there appears to be a codegen issue with
Bool8 which causes it to mangle values after the accessed range.
Conceptually, bool is the correc type to use in these places anyways,
and because the marshaller is not involved here, there are no problems
related to the marshaller's native-managed size mismatch.