Fix non '\0' terminated labels for InputText
Previously only the length of the label was allocated. As .net is
initializing arrays with 0 and (at least for this stackalloc here)
this happens in 4-byte chunks only strings with a length multiple of 4
were affected as in this case the allocated size exactly matched the
length of the string and so no 0 was left to be used for string end.
Now the array is generated always one byte larger than the label length,
so there will always be a 0 at the end.