Linked the src folder in readme to show the location of examples
I spent over 6 hours of dumb searching over the internet and debugging, just because I assumed the only available example is at the new-demo branch. I am running Monogame, so this example is far from perfect for my case. I was expecting the library code to be in the src/, and completely disregarded the directory. While it's definitely my fault, I believe that this change may help other absent minded individuals.